Clean up bogosities in use of random(3) and srandom(3) --- do not assume that RAND_MAX applies to them, since it doesn't.  Instead add a config.h parameter MAX_RANDOM_VALUE.  This is currently set at 2^31-1 but could be auto-configured if that ever proves necessary.  Also fix some outright bugs like calling srand() where srandom() is appropriate.

+++ src/backend/optimizer/geqo/geqo_main.c
@@ -106,9 +106,9 @@ geqo(Query *root)
 /* seed random number generator */
 /* XXX why is this done every time around? */
     if (Geqo_random_seed >= 0)
-        srandom(Geqo_random_seed);
+        srandom((unsigned int) Geqo_random_seed);
     else
-        srandom(time(NULL));
+        srandom((unsigned int) time(NULL));
 
 /* allocate genetic pool memory */
 	pool = alloc_pool(pool_size, number_of_rels);
